+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [5.0.0] - Jan 19, 2026
+
+### Changed
+
+- Use the official Spotify Web API for data
+
+## [4.2.1] - Oct 18, 2025
+
+### Changed
+
+- Change the latest TOTP secret endpoint
+
 ## [4.2.0] - Sep 27, 2025
 
 ### Added